backend/requirements.txt
affine==2.3.0
attrs==21.4.0
bcrypt==3.2.0
beautifultable==1.0.1
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cryptography==36.0.1
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
GDAL @ file:///C:/Users/Lenovo/Downloads/GDAL-3.4.1-cp310-cp310-win_amd64.whl
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.15.0
numpy==1.22.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2021.3
rasterio @ file:///C:/Users/Lenovo/Downloads/rasterio-1.2.10-cp310-cp310-win_amd64.whl
six==1.16.0
snuggs==1.4.7
SQLAlchemy==1.4.31
wcwidth==0.2.5
Werkzeug==2.0.3
